msm: display: pp: histogram collection for mdss
MSMFB_HISTOGRAM_START, MSMFB_HISTOGRAM_STOP and MSMFB_HISTOGRAM support
Only read the histogram data when there is a new display commit
Only generate the next histogram collection when there is a display commit
Cache the histogram data in order to respond quickly

+++ include/uapi/linux/msm_mdp.h
@@ -393,7 +393,7 @@ struct mdp_histogram_start_req {
 
 struct mdp_histogram_data {
 	uint32_t block;
-	uint8_t bin_cnt;
+	uint32_t bin_cnt;
 	uint32_t *c0;
 	uint32_t *c1;
 	uint32_t *c2;
